+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.3.4
 - Cube cut now cuts faces that are coplanar with the cube if they point outward
+- Even more sensible texturing on currently blank textured faces
 
 ## 1.3.3
 - Fix incorrect UV on extruded face when using the extrusion tool