+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.3.4
 - Cube cut now cuts faces that are coplanar with the cube if they point outward
 - Even more sensible texturing on currently blank textured faces
+- Fix a class of bugs relating to invalid cache on undo / redo
 
 ## 1.3.3
 - Fix incorrect UV on extruded face when using the extrusion tool